SW-213 Deletion of event shows and venues fixed All cases are supported. Refactoring.

+++ Server/SmartWalk.Server/Controllers/Base/EntityBaseController.cs
@@ -91,6 +91,9 @@ namespace SmartWalk.Server.Controllers.Base
         {
             if (CurrentSmartWalkUser == null) return new HttpUnauthorizedResult();
 
+            var access = _entityService.GetEntityAccess(CurrentSmartWalkUser.Record, entityId);
+            if (access != AccessType.AllowEdit) return new HttpUnauthorizedResult();
+
             _entityService.DeleteEntity(entityId);
 
             return RedirectToAction("List");
